% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_codes_lookup.R
\name{api_codes_lookup}
\alias{api_codes_lookup}
\title{api_codes_lookup (operational function)}
\usage{
api_codes_lookup()
}
\description{
This function is called to produce a simple API response codes lookup table.
}
