% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_end_point_map_df.R
\name{api_end_point_map_df}
\alias{api_end_point_map_df}
\title{api_end_point_map_df}
\usage{
api_end_point_map_df()
}
\value{
A data.frame with a map between api_end_point_type and end_point string.
}
\description{
api_end_point_map_df
}
