% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\docType{data}
\name{companies}
\alias{companies}
\title{Companies Data-Frame}
\format{An object of class tbl_df, tbl, data.frame.}
\usage{
companies
}
\description{
This dataset contains 1000 companies records from Company House. This extrct should be eused for illustrative purposes and it was collected \href{http://download.companieshouse.gov.uk/en_output.html}{here}.
}
\keyword{datasets}
