% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies_profile_collect.R
\name{companies_profile_collect}
\alias{companies_profile_collect}
\title{Companies House - companies_profile_collect}
\usage{
companies_profile_collect(companies_df, auth_api_key, time_to_rest = 3)
}
\arguments{
\item{companies_df}{A data-frame with CompanyNumber column (see companies data file as example).}

\item{auth_api_key}{The CompaniesHouse API Key. To generate one please go to \href{https://developer.companieshouse.gov.uk/api/docs/index/gettingStarted/apikey_authorisation.html}{CompaniesHouse Developers Page}.}

\item{time_to_rest}{In case the API repsonse is 429 ('Too Many Requests') the user can select the period to wait until it tries again. It defaults to 3 seconds.}
}
\value{
A list with two elements:
\itemize{
 \item{\strong{results_df}} {A data-frame with the results returned from the 'companies/ end point of the CompaniesHouse API. The results can be explored via \href{https://developer.companieshouse.gov.uk/api/docs/company/company_number/companyProfile-resource.html}{this link}.}
 \item{\strong{errorLogs}}  {In case some CompanyNumbers failed to return values either because of errors or because there was no results to be returned, these can be explored via errorLogs.}
}
}
\description{
This function is a wrapper for the Charges API from Companies House. Allows the user to collect Charges data by parsing a company number.
}
\details{
All Companies House APIs are accessed by company number. The user can collect company numbers from monthly extracts offered by CompanyHouse that can be accessed \href{http://download.companieshouse.gov.uk/en_output.html}{here}.
}
\examples{
\dontrun{
companies_api_collect(companies_df, auth_api_key , time_to_rest = 3)
}
}
