% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFs_from_DF.R
\name{DFs_from_DF}
\alias{DFs_from_DF}
\title{DFs_from_DF}
\usage{
DFs_from_DF(json_list)
}
\arguments{
\item{json_list}{A json items api response.}
}
\value{
A  dataframe with simplified response.
}
\description{
This function simplifies items returns that include DataFrames and lists as columns.
}
\examples{
\dontrun{
DFs_from_DF(df_items)
}
}
