% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromJSONtoDF_companies_base.R
\name{fromJSONtoDF}
\alias{fromJSONtoDF}
\title{fromJSONtoDF}
\usage{
fromJSONtoDF(json_list, api_end_point, company_number)
}
\arguments{
\item{json_list}{A list representing the JSON response from the companies/ (base) API end-point.}

\item{api_end_point}{A character value describing which end-point to call.}

\item{company_number}{Company number}
}
\value{
A  dataframe with a single row response.
}
\description{
This function is simplfies the conversion of the JSON response from the companies (base) API end-point into a data-frame format.
}
\examples{
\dontrun{
fromJSONtoDF(json_list)
}
}
