% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromJSONtoDF_companies_base.R
\name{fromJSONtoDF_companies_base}
\alias{fromJSONtoDF_companies_base}
\title{fromJSONtoDF_companies_base}
\usage{
fromJSONtoDF_companies_base(json_list)
}
\arguments{
\item{json_list}{A list representing the JSON response from the companies/ (base) API end-point.}
}
\value{
A  dataframe with a single row response.
}
\description{
This function is simplfies the conversion of the JSON response from the companies (base) API end-point into a data-frame format.
}
\examples{
\dontrun{
fromJSONtoDF_companies_base(json_list)
}
}
